End of Timer Behavior
Added a few comments for the changes I made last night then added the end of the game timer behavior. Once the timer runs out an unbeatable fleet of enemies will spawn in to destroy the player. They will have a few seconds to eventually jump and/or gather supplies before being killed.